Visualizations made in Tableau
The full dashboard can be seen here. I also explored relationships between weight and muscle and found them to be so correlated that I didn't find it interesting to pursue. Similarly BMI is a direct function of weight, so not worth exploring as all things that BMI tells me can be equally found through weight. Most curious is the negative trend between calories and weight. Quick reflection lead me to the obvious conclusion: I weigh myself in the morning and track food later in the day, so a better relationship between tables would be calories from the day before linked to weight of the current day.

Unfortunately Tableau Public wouldn't allow me to make a relationship to a calculated field, so I couldn't enact the one day lag between calories and weight. I'll do that in my more in depth Python analysis. Also I was unable to get the animations for a line chart of weight vs calories to work well in Tableau Public. Internet search showed this animation difficulty is a common problem on Tableau Public, alas. Tableau animations can be so cool!

Url for full Tableau Public dashboard
 https://public.tableau.com/views/PersonalHealthProject/WeightNutritionSummaryDashboard?:language=en-US&:display_count=n&:origin=viz_share_link